:sparkles: FixConfig
